fix: Resolve build errors in monthly budget feature
This commit fixes the build errors that occurred after the initial implementation of the monthly budget feature.

- **Repository Package Revert:** Reverted the repository package name back to `com.example.finanzas.data.repository` to resolve Hilt dependency injection failures. Updated all import statements across ViewModels, UseCases, and Hilt modules.
- **Hilt Dependency Injection:** Ensured `RepositoryModule.kt` correctly provides `BudgetDao` to the `FinanzasRepositoryImpl` constructor.
- **DAO Type Mismatch:** Corrected a type mismatch in `BudgetDao` by changing the `budgetId` in the `BudgetCategory` entity to a `Long`.
- **UI Placeholder:** Implemented the category icon in the `BudgetScreen` that was previously commented out.

+++ app/src/main/java/com/example/finanzas/data/local/dao/BudgetDao.kt
@@ -29,7 +29,7 @@ interface BudgetDao {
     @Transaction
     suspend fun saveBudget(budget: Budget, categories: List<BudgetCategory>) {
         val budgetId = insertBudget(budget)
-        val categoriesWithId = categories.map { it.copy(budgetId = budgetId.toInt()) }
+        val categoriesWithId = categories.map { it.copy(budgetId = budgetId) }
         insertBudgetCategories(categoriesWithId)
     }
 }
